myblogsite/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-taggit==4.0.0
Markdown==3.4.3
psycopg2-binary==2.9.6
sqlparse==0.4.4
